+++ Tests/GRDBTests/AssociationPrefetchingSQLTests.swift
@@ -1898,43 +1898,6 @@ class AssociationPrefetchingSQLTests: GRDBTestCase {
         }
     }
     
-    func testLimit() throws {
-        struct Player: TableRecord { }
-        struct Team: TableRecord {
-            static let players = hasMany(Player.self)
-        }
-        
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.write { db in
-            try db.create(table: "team") { t in
-                t.autoIncrementedPrimaryKey("teamId")
-            }
-            try db.create(table: "player") { t in
-                t.autoIncrementedPrimaryKey("playerId")
-                t.column("parentId", .integer).references("team")
-            }
-            try db.execute(sql: "INSERT INTO team DEFAULT VALUES")
-            
-            do {
-                sqlQueries.removeAll()
-                let association = Team.players.limit(10, offset: 5)
-                let request = Team.including(all: association)
-                _ = try Row.fetchAll(db, request)
-                let selectQueries = sqlQueries.filter(isSelectQuery)
-                XCTAssertEqual(selectQueries, [
-                    """
-                    SELECT * FROM "team"
-                    """,
-                    """
-                    SELECT *, "parentId" AS "grdb_parentId" \
-                    FROM "player" \
-                    WHERE "parentId" = 1 \
-                    LIMIT 10 OFFSET 5
-                    """])
-            }
-        }
-    }
-    
     func testCTE() throws {
         struct Player: TableRecord { }
         struct Team: TableRecord {